refactor: rename "Graviti", "schema", "utility.user" and "utility.lazy2"
- `Graviti` -> `Platform`

- `schema` -> `portex`

- `utility.user` -> `utility.collections`

- `utility.lazy2` -> `utility.paging`

PR Closed: https://github.com/Graviti-AI/graviti-python-sdk/pull/143
