+++ graviti/utility/collections.py
